Prepend rather than append the extra options to argv
argv contains any options for the node process, followed eventually by the script to execute
and then the options for the script. Adding to the end therefore would give the option to the script
and not the NodeJS process.

Factor out more logic from cli.ts and add tests for these cases