Use a configurable fraction of the detected memory limit for the `--max-old-space-size` parameter
The memory limit that the cgroups interface exposes refers to the complete process memory, of which
the old space heap is only a part. There doesn't seem to be an obvious way to make NodeJS work out
a good sizing based on a provided maximum value, so for now we just simply apply a heuristic, and
use 70% of the value.

This can be configured by
1. Setting the environment variable CGROUP_MEMORY_FRACTION to a value (ideally between 0 and 1 ^^)
2. Providing an additional `--cgroup-memory-fraction[=value| value]` argument, followed by a mandatory
   '--'